moved vt base mixin to global mixins folder
