+++ api_app/analyzers_manager/observable_analyzers/vt/vt3_get.py
@@ -2,9 +2,8 @@
 # See the file 'LICENSE' for copying permission.
 
 from api_app.analyzers_manager.classes import ObservableAnalyzer
+from api_app.mixins import VirusTotalv3AnalyzerMixin
 from tests.mock_utils import MockUpResponse, if_mock_connections, patch
-
-from .vt3_base import VirusTotalv3AnalyzerMixin
 
 
 class VirusTotalv3(ObservableAnalyzer, VirusTotalv3AnalyzerMixin):
