+++ api_app/analyzers_manager/observable_analyzers/vt/vt3_intelligence_search.py
@@ -5,10 +5,10 @@ from typing import Dict
 import requests
 
 from api_app.analyzers_manager.classes import ObservableAnalyzer
+from api_app.mixins import VirusTotalv3AnalyzerMixin
 from tests.mock_utils import MockUpResponse, if_mock_connections, patch
 
 from ...exceptions import AnalyzerRunException
-from .vt3_base import VirusTotalv3AnalyzerMixin
 
 
 class VirusTotalv3Intelligence(ObservableAnalyzer, VirusTotalv3AnalyzerMixin):
